Bag supports adding and resolving references out of bound (#52443)
Part of #52133

This is a refactoring of the Bag implementation.

Before: resolution of references is like a functional constructor. We
take a number of references, and try to resolve all of them. The result
is a Bag that can be queried for these and associated references.

This works well for search filtering `f:has.owner(...)` since everything
hinges on the query text in the predicate.

For graphQL API owners unification (bunch of owners of different data
kinds), and `select:file.owners` search, the model is different. Many
different owner references appear, and they need to be reconciled with
each other (in an online fashion for search).

So the implementation changes to the following set of bag methods:

- `add` puts one more potentially unresolved reference, like `{handle:
"foo"}`, `{email:"bar@example.com"}` or `{userID: 42}`
- `resolve` takes in all the references that were `add`-ed and not
attempted to be resolved before
  - it tries to find users associated with these references
- if these users are found, and are not in the bag, they are
`augment`-ed, that is
- all the other references for given user are fetched from the database
    - and linked back (`linkBack`) into the bag
- `contains` can look up resolved and unresolved entities based on the
references above.

## Test plan

Existing unit tests - this is a refactoring