runMochaTestsWithPercy: preserve args
If an arg is passed to runMochaTestsWithPercy that contains a space, it is split into 2 args when the underlying command is run. This causes problems with (eg) `-f 'some test'`, which is used to filter Mocha tests to run, where only `some` is used and `test` is treated as a separate command-line arg.

This does not cause any known problems today, but it is a bug and makes it hard to run a subset of integration tests through this runner in local dev.